backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "authlib>=1.6.6",
    "bcrypt>=5.0.0",
    "boto3>=1.42.38",
    "alembic>=1.13.1",
    "fastapi[standard]>=0.128.0",
    "langchain>=1.2.7",
    "langchain-groq>=1.1.1",
    "langchain-mistralai>=1.1.1",
    "langchain-postgres>=0.0.16",
    "langchain-redis>=0.2.5",
    "langgraph>=1.0.7",
    "logfire[fastapi,httpx,requests,sqlalchemy,system-metrics]>=4.21.0",
    "loguru>=0.7.3",
    "pip-system-certs>=5.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "pymupdf>=1.24.9",
]
